Complete 1.21.8 protocol implementation  - changed reflection library to own tooling, new reflect_map tool to parse headers of structs and provide header for iterating structs and serializing enums(a lot faster and less bugs)  - change new_packets api name to packets  - partial implementation for selector  - added std::hash for id_source  - added api::permissions::is_in_group function  - moved all api cpp files to api/bin  - SharedClientData now directly stores compatible_plugins instead of names  - changed std::vector in api::plugin to list_array  - added new bounds: cubic_bounds_block_radius_out, spherical_bounds_chunk_out and spherical_bounds_block_out

+++ src/util/readers.hpp
@@ -297,7 +297,7 @@ namespace copper_server {
     }
 
     static void WriteString(list_array<uint8_t>& data, const std::string& str, int32_t max_string_len = INT32_MAX) {
-        if (str.size() > max_string_len)
+        if (str.size() > (size_t)max_string_len)
             throw std::out_of_range("actual string len out of range");
         WriteVar<int32_t>(str.size(), data);
         data.push_back((uint8_t*)str.data(), str.size());
